+++ source/platforms/git/localGetDiff.ts
@@ -2,10 +2,11 @@ import { debug } from "../../debug"
 import { spawn } from "child_process"
 
 const d = debug("localGetDiff")
-
-export const localGetDiff = (base: string, head: string) =>
+const useCommittedDiffArgs = (base: string, head: string) => ["diff", `${base}...${head}`]
+const useStagingChanges = (base: string) => ["diff", base]
+export const localGetDiff = (base: string, head: string, staging: boolean = false) =>
   new Promise<string>(done => {
-    const args = ["diff", `${base}...${head}`]
+    const args = staging ? useStagingChanges(base) : useCommittedDiffArgs(base, head)
     let stdout = ""
 
     const child = spawn("git", args, { env: process.env })